Renamed bible_screen to apps_screen and added "Open Bible" button. Added dividers to HomePage.
